Bug fix: preserve cluster compression type after clear().
By nulling out the cluster implementation on clear, we were also
resetting the cluster compression type.  This caused the ZimCreator
API to not compress compressible clusters after the first one.
This is a follow-up to f5de40f94b30795f42bb9388cbb46df9cd605167,
which reused cluster objects instead of recreating them from scratch
each time.

Related issue: the `size()` of an empty cluster is actually 4, not 0,
since even an empty cluster contains an field which counts the number
of offsets in the cluster, and change the ClusterImpl initializer so
the compresion type doesn't change from zimcompNone to zimcompDefault
as soon as the ClusterImpl is created.

Change-Id: I468a1719a33c450db9a28d9704b539bdb97cd7fc