Fix memory leak in zimwriterfs.
We are using polymorphism in articlesource.cpp (a Article* pointer but
derived class as real instance).
If we want to correctly delete allocated resources, we need to let compiler
get the correct destructor in the vtable.

In our case, the data member of FileArticle was never destructed and
the internal buffer of the std::string was leaked (even if the article was
correctly deleted).

Change-Id: I58833f48a440e9c91a97637d6c11e5642eaf64ce