Fix use of ENABLE_* by preprocessor.
If a compression lib is not present, the associated ENABLE_* is not
defined.
So we need to always test for the definition existance and not for its
value.
This is not the case in zimcreator.cpp with the #ifdef/#elif construction.

We also change other (correct) syntax #ifdef to #if defined() to keep some
consistency.

Change-Id: I86d4309bfcdeeb3356d0fb4f192d0849a5e57275